requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.31.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
tenacity==9.0.0
orjson==3.10.7
loguru==0.7.2
typing-extensions>=4.12.2
# Toggle-able
langchain==0.3.7
langchain-openai==0.2.6

# Optional tools used by the chat pipeline
python-dateutil>=2.9.0
